requirements.txt
Django>=3.1.5,<3.2
djangorestframework>=3.12.2,<3.13.0
django-filter 2.4.0